fix: ディレクトリ構成の変更
