+++ api/src/infra/xml/xml2js/badge.ts
@@ -1,5 +1,5 @@
 import type { Badge } from "@/entity/badge";
-import { XmlCore } from "@/infra/xml/core";
+import { XmlCore } from "@/infra/xml/xml2js/core";
 
 export type BadgeXml = {
   $: {
