+++ api/src/infra/xml/xml2js/tag.ts
@@ -1,5 +1,5 @@
 import type { Tag } from "@/entity/tag";
-import { XmlCore } from "@/infra/xml/core";
+import { XmlCore } from "@/infra/xml/xml2js/core";
 
 export type TagXml = {
   $: {
