+++ api/src/infra/xml/xml2js/user.ts
@@ -1,5 +1,5 @@
 import type { User } from "@/entity/user";
-import { XmlCore } from "@/infra/xml/core";
+import { XmlCore } from "@/infra/xml/xml2js/core";
 
 export type UserXml = {
   $: {
