+++ api/src/infra/xml/xml2js/vote.ts
@@ -1,5 +1,5 @@
 import type { Vote } from "@/entity/vote";
-import { XmlCore } from "@/infra/xml/core";
+import { XmlCore } from "@/infra/xml/xml2js/core";
 
 export type VoteXml = {
   $: {
